Se añade columna price en tabla de cars

+++ db/migrate/20250924025412_add_price_to_cars.rb
@@ -1,5 +1,5 @@
 class AddPriceToCars < ActiveRecord::Migration[7.1]
   def change
-    add_column :cars, :price, :float
+    add_column :cars, :price, :integer
   end
 end
